+++ skbio/alignment/tests/test_alignment.py
@@ -16,6 +16,7 @@ except ImportError:  # python3 system
     from io import StringIO
 import tempfile
 
+import matplotlib as mpl
 import numpy as np
 from scipy.spatial.distance import hamming
 
@@ -714,6 +715,35 @@ class AlignmentTests(TestCase):
         self.assertEqual(self.a1.sequence_length(), 13)
         self.assertEqual(self.a2.sequence_length(), 5)
         self.assertEqual(self.empty.sequence_length(), 0)
+
+    def test_sequence_logo_error(self):
+        a1 = Alignment([DNA('ATTGK', id='seq1'),
+                        DNA('ACGTT', id='seq2'),
+                        DNA('CTCTG', id='seq3')])
+        with self.assertRaises(AlignmentError):
+            a1.sequence_logo()
+        a2 = Alignment([])
+        with self.assertRaises(AlignmentError):
+            a2.sequence_logo()
+
+    def test_sequence_logo_mixed_gap_chars_and_single_seq(self):
+        a1 = Alignment([DNA('A-T.C', id='seq1')])
+        fig = Alignment.sequence_logo(a1)
+        self.check_sequence_logo_figure_sanity(fig, ['1', '2', '3', '4', '5'])
+
+    def check_sequence_logo_figure_sanity(self, fig, xticks):
+        self.assertIsInstance(fig, mpl.figure.Figure)
+        axes = fig.get_axes()
+        ax = axes[0]
+        x = []
+        for tick in ax.get_xticklabels():
+            x.append(tick.get_text())
+        self.assertEqual(x, xticks)
+        y = []
+        for tick in ax.get_yticklabels():
+            y.append(tick.get_text())
+        self.assertEqual(y[0], '0.0')
+        self.assertEqual(y[-1], '1.0')
 
     def test_validate_lengths(self):
         self.assertTrue(self.a1._validate_lengths())